added graphQL schemas for updateProduct mutation and new types and hooks generated by codegen

+++ client/src/schema/queries.ts
@@ -21,7 +21,7 @@ export const GET_PRODUCT = gql`
 `;
 
 export const GET_PRODUCT_BY_ID = gql`
-  query getProductById($getProductByIdId: Float!) {
+  query getProductById($getProductByIdId: Int!) {
     getProductById(id: $getProductByIdId) {
       id
       reference
